test(material/input): combine shared tests
Since we only have one module, we don't need separate shared tests anymore. Also includes one list cleanup for a file that doesn't exist.
